looking to save checked button
